buscador funcional

+++ app/src/main/java/com/example/trabajo_final/Anuncio.kt
@@ -15,7 +15,7 @@ data class Anuncio(
     var usuarioPaseador: String? = null,
     var estado: String? = null,
     var tipoAnuncio: String? = null,
-    var precio: Float? = null,
+    var precio: Double? = null,
     var idmascota: List<String>? = null,
     var nombreMascota: List<String>? = null,
     var razaMascota: List<String>? = null,
